requirements.txt
python-telegram-bot==20.3
requests>=2.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
stable-baselines3==2.0.0
torch==2.0.1
transformers==4.31.0
beautifulsoup4==4.12.2
tweepy==4.14.0
gym==0.26.2
coinbase-advanced-trade==0.3.2
